pyproject.toml
[tool.poetry]
name = "restapi"
version = "0.1.0"
description = ""
authors = ["AusAura <91804331+AusAura@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
sqlalchemy = "^2.0.22"
psycopg2 = "^2.9.9"
alembic = "^1.12.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.6"
fastapi-mail = "^1.4.1"
redis = "4.2"
pydantic-settings = "^2.0.3"
fastapi-limiter = "^0.1.5"
cloudinary = "^1.36.0"
pytest = "^7.4.3"
httpx = "^0.25.1"

[tool.poetry.group.dev.dependencies]
icecream = "^2.1.3"
sphinx = "^7.2.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
pythonpath = [
  "."
]